+++ tests/contrib/django/test_basic.py
@@ -7,7 +7,8 @@ from uwsgiconf.contrib.django.uwsgify.models import Task
 
 
 @pytest.mark.skip("Live server for development")
-def test_run_app(run_app):
+def test_run_app(run_app, settings):
     now = timezone.now()
     Task.register("task_1", dt_acquired=now-timedelta(hours=2), dt_released=now-timedelta(minutes=35))
+    Task.register("task_2", dt_acquired=now-timedelta(hours=1), dt_released=now-timedelta(hours=8))
     run_app()